src/requirements.txt
Flask>=3.0.2
Pillow>=10.0.0
Pillow>=10.2.0
python-dotenv>=1.0.1
pywin32>=306
